\name{SC19039-package}
\alias{SC19039-package}
\docType{package}
\title{
  NMF and SNMF
}
\description{
  Realized standard non-negative matrix factorization and sparse non-negative matrix factorization
}
\details{
  The wfNMF function decomposes a matrix V into the product of two non-negative matrices: V = W * H. The SNMF function further controls the number of non-zero elements in H.
}
\author{
Fan Wu.

Maintainer: Fan Wu<wufan25@mail.ustc.edu.cn>
}
\keyword{NMF ,BeSS,SNMF }
